config: 모듈 분리
