+++ cafe-admin/src/main/java/kr/co/yapp/cafe/controller/IndexController.java
@@ -1,0 +1,12 @@
+package kr.co.yapp.cafe.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class IndexController {
+    @GetMapping
+    public String index() {
+        return "index";
+    }
+}
