+++ cafe-api/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    implementation project(':cafe-domain')
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
