+++ cafe-batch/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    implementation project(':cafe-domain')
+    implementation 'org.springframework.boot:spring-boot-starter-batch'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
