+++ cafe-domain/build.gradle
@@ -1,0 +1,12 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    runtimeOnly 'com.h2database:h2'
+}
+
+jar {
+    enabled = true
+}
+
+bootJar {
+    enabled = false
+}
