+++ cafe-domain/src/test/java/kr/co/yapp/cafe/CafeApplicationTests.java
@@ -1,0 +1,11 @@
+package kr.co.yapp.cafe;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CafeApplicationTests {
+    @Test
+    void contextLoads() {
+    }
+}
