+++ cafe-domain/src/test/java/kr/co/yapp/cafe/TestCafeApplication.java
@@ -1,0 +1,7 @@
+package kr.co.yapp.cafe;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestCafeApplication {
+}
