+++ settings.gradle
@@ -1,1 +1,6 @@
 rootProject.name = 'cafe'
+include 'cafe-domain'
+include 'cafe-api'
+include 'cafe-batch'
+include 'cafe-admin'
+
