--- src/main/java/kr/co/yapp/cafe/CafeApplication.java
@@ -1,13 +1,0 @@
-package kr.co.yapp.cafe;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CafeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CafeApplication.class, args);
-	}
-
-}
