--- src/test/java/kr/co/yapp/cafe/CafeApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.co.yapp.cafe;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CafeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
